+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+-   Support status property on events [#422](https://github.com/markuspoerschke/iCal/pull/421)
 -   Support categories property on events [#421](https://github.com/markuspoerschke/iCal/pull/421)
 
 ## [2.5.1] - 2022-04-26